Implemented TCPHeader class, with encoding and decoding
Just data transfer and keeping track of Seq and Ack numbers for now, no
congestion control. Sequence numbers currently cout up all the way to
2^16 (the field is uint16_t), whereupon they wrap 0. They probably need
to wrap the max seq number field but we'll figure that out later. Simple
text file content is coming over the wire correctly, but when I run it
on the cat image it hangs.